Move previewformat package from subpackage under post
