+++ api/src/main/java/com/percolate/sdk/api/request/previewformat/PreviewFormatRequest.java
@@ -1,4 +1,4 @@
-package com.percolate.sdk.api.request.post.previewformat;
+package com.percolate.sdk.api.request.previewformat;
 
 import com.percolate.sdk.api.PercolateApi;
 import com.percolate.sdk.api.utils.RetrofitApiFactory;
